ДЗ 2.4 Исключения.Распр.методы по классам
